Add non working initial zonal statistics code
