+++ .github/workflows/deploy.yml
@@ -41,6 +41,15 @@ jobs:
       - name: Build and push Docker image
         uses: docker/build-push-action@ad44023a93711e3deb337508980b4b5e9bcdc5dc
         with:
+          context: ./pygeoapi
+          push: true
+          tags: |
+            ghcr.io/${{ github.repository_owner }}/pygeoapi:latest
+            ghcr.io/${{ github.repository_owner }}/pygeoapi:${{ steps.timestamp.outputs.timestamp }}
+
+      - name: Build and push Docker image
+        uses: docker/build-push-action@ad44023a93711e3deb337508980b4b5e9bcdc5dc
+        with:
           context: ./worker/error-handler
           push: true
           tags: |
